[ContentBundle] add custom config for ORM in ContentBundle (#165)
* [ContentBundle] add custom config for ORM in ContentBundle
* [ContentBundle] add support for articles and routes in ORM
* store tenants and orgs using doctrine orm
* added orm factory
* separate article factory
* added specs for article factories
* improve theme resolver
* pagination changes
* add static prefix generation to route service
* rebuild article media loader
* implement tree for routes
* add renditions to article media
* fix more tests
* fix route collection tests
* move image renditions creation to media factory
* fix MetaRouter test
* move menu to orm, MenuBundle
* code style and organization fixes
* fixed menu tests
* added missing param
* make article and route tenant aware
* disable phpcr, (SWP-336) Make menu triggering template flexible
* fixed menu bundle mapping
* Swp 330 (#34)
* Add domain field to Tenant class, fix providers specs
* refactor menu bundle a bit
* added tenant spec
* revert route enhancer changes, fix specs
* expose tenant domain
* revert route enhancer changes, added text type field for article metadata
* update vendors, add timestampable to Article Media
* update changelog

+++ src/SWP/Bundle/ContentBundle/Loader/ArticleLoader.php
@@ -16,10 +16,10 @@ declare(strict_types=1);
 
 namespace SWP\Bundle\ContentBundle\Loader;
 
-use Jackalope\Query\SqlQuery;
-use SWP\Bundle\ContentBundle\Doctrine\ODM\PHPCR\Article;
-use SWP\Bundle\ContentBundle\Doctrine\ODM\PHPCR\ArticleInterface;
-use SWP\Bundle\ContentBundle\Doctrine\ODM\PHPCR\Route;
+use Doctrine\Common\Persistence\ObjectManager;
+use SWP\Component\Common\Criteria\Criteria;
+use SWP\Bundle\ContentBundle\Model\ArticleInterface;
+use SWP\Bundle\ContentBundle\Model\RouteInterface;
 use SWP\Bundle\ContentBundle\Provider\ArticleProviderInterface;
 use SWP\Bundle\ContentBundle\Provider\RouteProviderInterface;
 use SWP\Component\TemplatesSystem\Gimme\Context\Context;
@@ -27,12 +27,12 @@ use SWP\Component\TemplatesSystem\Gimme\Factory\MetaFactoryInterface;
 use SWP\Component\TemplatesSystem\Gimme\Loader\LoaderInterface;
 use SWP\Component\TemplatesSystem\Gimme\Meta\Meta;
 use SWP\Component\TemplatesSystem\Gimme\Meta\MetaCollection;
-use Doctrine\ODM\PHPCR\DocumentManager;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 /**
  * Class ArticleLoader.
  */
-class ArticleLoader implements LoaderInterface
+class ArticleLoader extends PaginatedLoader implements LoaderInterface
 {
     /**
      * @var ArticleProviderInterface
@@ -45,7 +45,7 @@ class ArticleLoader implements LoaderInterface
     protected $routeProvider;
 
     /**
-     * @var DocumentManager
+     * @var ObjectManager
      */
     protected $dm;
 
@@ -69,14 +69,14 @@ class ArticleLoader implements LoaderInterface
      *
      * @param ArticleProviderInterface $articleProvider
      * @param RouteProviderInterface   $routeProvider
-     * @param DocumentManager          $dm
+     * @param ObjectManager            $dm
      * @param MetaFactoryInterface     $metaFactory
      * @param Context                  $context
      */
     public function __construct(
         ArticleProviderInterface $articleProvider,
         RouteProviderInterface $routeProvider,
-        DocumentManager $dm,
+        ObjectManager $dm,
         MetaFactoryInterface $metaFactory,
         Context $context
     ) {
@@ -109,62 +109,56 @@ class ArticleLoader implements LoaderInterface
      */
     public function load($type, $parameters = [], $responseType = LoaderInterface::SINGLE)
     {
-        $article = null;
-
+        $criteria = new Criteria();
         if ($responseType === LoaderInterface::SINGLE) {
-            if (array_key_exists('contentPath', $parameters)) {
-                $article = $this->dm->find('SWP\Bundle\ContentBundle\Doctrine\ODM\PHPCR\Article', $parameters['contentPath']);
-                if (null !== $article && !$article->isPublished()) {
-                    $article = null;
-                }
-            } elseif (array_key_exists('article', $parameters)) {
+            if (array_key_exists('article', $parameters) && $parameters['article'] instanceof ArticleInterface) {
                 $this->dm->detach($parameters['article']);
-                $article = $this->dm->find('SWP\Bundle\ContentBundle\Doctrine\ODM\PHPCR\Article', $parameters['article']->getId());
-                if (null !== $article && !$article->isPublished()) {
-                    $article = null;
-                }
+                $criteria->set('id', $parameters['article']->getId());
             } elseif (array_key_exists('slug', $parameters)) {
-                $article = $this->dm->getRepository('SWP\Bundle\ContentBundle\Doctrine\ODM\PHPCR\Article')->findOneBy([
-                    'slug' => $parameters['slug'],
-                    'status' => ArticleInterface::STATUS_PUBLISHED,
-                ]);
+                $criteria->set('slug', $parameters['slug']);
             }
 
-            return $this->getArticleMeta($article);
+            try {
+                return $this->getArticleMeta($this->articleProvider->getOneByCriteria($criteria));
+            } catch (NotFoundHttpException $e) {
+                return;
+            }
         } elseif ($responseType === LoaderInterface::COLLECTION) {
-            $route = null;
             $currentPage = $this->context->getCurrentPage();
-            if (array_key_exists('route', $parameters)) {
-                if (null !== $currentPage && $currentPage->getValues()->getId() === $parameters['route']) {
-                    $route = $currentPage->getValues();
-                } else {
-                    $route = $this->routeProvider->getOneById($parameters['route']);
-                }
-            } elseif (null !== $currentPage) {
+            $route = null;
+
+            if (null !== $currentPage) {
                 $route = $currentPage->getValues();
             }
 
-            if (null !== $route && is_object($route)) {
+            if (array_key_exists('route', $parameters)) {
+                if (null === $route || ($route instanceof RouteInterface && $route->getId() !== $parameters['route'])) {
+                    if (is_int($parameters['route'])) {
+                        $route = $this->routeProvider->getOneById($parameters['route']);
+                    } elseif (is_string($parameters['route'])) {
+                        $route = $this->routeProvider->getOneByStaticPrefix($parameters['route']);
+                    }
+                }
+            }
+
+            if ($route instanceof RouteInterface) {
+                $criteria->set('route', $route);
+            } else {
+                return;
+            }
+
+            $criteria = $this->applyPaginationToCriteria($criteria, $parameters);
+            $articles = $this->articleProvider->getManyByCriteria($criteria);
+            if ($articles->count() > 0) {
                 $metaCollection = new MetaCollection();
-                $routeIdentifier = $this->dm->getNodeForDocument($route)->getIdentifier();
-                $metaCollection->setTotalItemsCount($this->getRouteArticlesQuery($routeIdentifier, [])->execute()->getRows()->count());
-                $query = $this->getRouteArticlesQuery($routeIdentifier, $parameters);
-
-                if (isset($parameters['limit'])) {
-                    $query->setLimit($parameters['limit']);
-                }
-
-                if (isset($parameters['start'])) {
-                    $query->setOffset($parameters['start']);
-                }
-
-                $articles = $this->dm->getDocumentsByPhpcrQuery($query, Article::class);
+                $metaCollection->setTotalItemsCount($this->articleProvider->getCountByCriteria($criteria));
                 foreach ($articles as $article) {
                     $articleMeta = $this->getArticleMeta($article);
                     if (null !== $articleMeta) {
                         $metaCollection->add($articleMeta);
                     }
                 }
+                unset($articles, $route, $criteria);
 
                 return $metaCollection;
             }
@@ -192,21 +186,5 @@ class ArticleLoader implements LoaderInterface
         }
 
         return;
-    }
-
-    /**
-     * @param string $routeIdentifier
-     * @param array  $parameters
-     *
-     * @return SqlQuery
-     */
-    private function getRouteArticlesQuery(string $routeIdentifier, array $parameters) : SqlQuery
-    {
-        $order = [];
-        if (array_key_exists('order', $parameters) && is_array($parameters['order'])) {
-            $order = $parameters['order'];
-        }
-
-        return $this->articleProvider->getRouteArticlesQuery($routeIdentifier, $order);
     }
 }
